+++ starsky/starskytest/starsky.foundation.readmeta/Services/ReadMeta_ExifReadTest.cs
@@ -16,6 +16,7 @@ using starsky.project.web.Attributes;
 using starskytest.FakeCreateAn;
 using starskytest.FakeCreateAn.CreateAnImageA330Raw;
 using starskytest.FakeCreateAn.CreateAnImageA6600Raw;
+using starskytest.FakeCreateAn.CreateAnImageA6700;
 using starskytest.FakeCreateAn.CreateAnImageLongDescriptionTitle;
 using starskytest.FakeCreateAn.CreateAnImageWebP;
 using starskytest.FakeCreateAn.CreateAnQuickTimeMp4;
@@ -332,7 +333,7 @@ public sealed class ExifReadTest
 		Assert.AreEqual(new DateTime(2018, 04, 22, 16, 14, 54, DateTimeKind.Local),
 			item.DateTime);
 
-		Assert.AreEqual("Sony|SLT-A58|24-105mm F3.5-4.5", item.MakeModel);
+		Assert.AreEqual("Sony|SLT-A58|24-105mm F3.5-4.5|", item.MakeModel);
 		Assert.AreEqual("Sony", item.Make);
 		Assert.AreEqual("SLT-A58", item.Model);
 		Assert.AreEqual("24-105mm F3.5-4.5", item.LensModel);
@@ -384,10 +385,47 @@ public sealed class ExifReadTest
 		Assert.AreEqual(new DateTime(2024, 11, 7, 14, 05, 35, DateTimeKind.Local),
 			item.DateTime);
 
-		Assert.AreEqual("Sony|ILCE-6600|E 18-200mm F3.5-6.3 OSS LE", item.MakeModel);
+		Assert.AreEqual("Sony|ILCE-6600|E 18-200mm F3.5-6.3 OSS LE|", item.MakeModel);
 		Assert.AreEqual("Sony", item.Make);
 		Assert.AreEqual("ILCE-6600", item.Model);
 		Assert.AreEqual("E 18-200mm F3.5-6.3 OSS LE", item.LensModel);
+		Assert.AreEqual(string.Empty, item.MakeCameraSerial);
+		Assert.AreEqual(ImageStabilisationType.On, item.ImageStabilisation);
+	}
+	
+	[TestMethod]
+	public void ExifRead_CreateAnImageA6700_ReadExifFromFileTest()
+	{
+		var newImage = new CreateAnImageA6700().Bytes.ToArray();
+		var fakeStorage = new FakeIStorage(new List<string> { "/" },
+			new List<string> { "/test.jpg" }, new List<byte[]> { newImage });
+
+		var item =
+			new ReadMetaExif(fakeStorage, null!, new FakeIWebLogger()).ReadExifFromFile(
+				"/test.jpg");
+
+		Assert.AreEqual(ColorClassParser.Color.None, item.ColorClass);
+		Assert.AreEqual(string.Empty, item.Description);
+		Assert.IsFalse(item.IsDirectory);
+		Assert.AreEqual("t", item.Tags);
+		Assert.AreEqual(string.Empty, item.Title);
+		Assert.AreEqual(52.1333222222, item.Latitude, 0.000001);
+		Assert.AreEqual(6.0067055555, item.Longitude, 0.000001);
+		Assert.AreEqual(1, item.ImageHeight);
+		Assert.AreEqual(1, item.ImageWidth);
+		Assert.AreEqual("Loenen", item.LocationCity);
+		Assert.AreEqual("Gelderland", item.LocationState);
+		Assert.AreEqual("Nederland", item.LocationCountry);
+		Assert.AreEqual(36, item.LocationAltitude);
+		Assert.AreEqual(31, item.FocalLength);
+		Assert.AreEqual(new DateTime(2025, 5, 15, 10, 29, 48, DateTimeKind.Local),
+			item.DateTime);
+
+		Assert.AreEqual("Sony|ILCE-6700|E 18-200mm F3.5-6.3 OSS LE|02059361", item.MakeModel);
+		Assert.AreEqual("Sony", item.Make);
+		Assert.AreEqual("ILCE-6700", item.Model);
+		Assert.AreEqual("E 18-200mm F3.5-6.3 OSS LE", item.LensModel);
+		Assert.AreEqual("02059361", item.MakeCameraSerial);
 		Assert.AreEqual(ImageStabilisationType.On, item.ImageStabilisation);
 	}
 
@@ -419,7 +457,7 @@ public sealed class ExifReadTest
 		Assert.AreEqual(new DateTime(2010, 2, 3, 17, 23, 7, DateTimeKind.Local),
 			item.DateTime);
 
-		Assert.AreEqual("Sony|DSLR-A330|Sony DT 18-55mm F3.5-5.6 SAM (SAL1855) or SAM II",
+		Assert.AreEqual("Sony|DSLR-A330|Sony DT 18-55mm F3.5-5.6 SAM (SAL1855) or SAM II|",
 			item.MakeModel);
 		Assert.AreEqual("Sony", item.Make);
 		Assert.AreEqual("DSLR-A330", item.Model);
@@ -459,8 +497,9 @@ public sealed class ExifReadTest
 		Assert.AreEqual(string.Empty, item.LocationState);
 		Assert.AreEqual(5.8325499999, item.Longitude, 0.0000000001);
 		Assert.AreEqual("Sony", item.Make);
-		Assert.AreEqual("Sony|ILCE-6600|E 18-200mm F3.5-6.3 OSS LE", item.MakeModel);
+		Assert.AreEqual("Sony|ILCE-6600|E 18-200mm F3.5-6.3 OSS LE|0", item.MakeModel);
 		Assert.AreEqual("ILCE-6600", item.Model);
+		Assert.AreEqual(string.Empty, item.MakeCameraSerial);
 		Assert.AreEqual(ImageRotation.Rotation.Horizontal, item.Orientation);
 		Assert.AreEqual("1/60", item.ShutterSpeed);
 		Assert.AreEqual(string.Empty, item.SidecarExtensions);