+++ history.md
@@ -45,6 +45,7 @@ Semantic Versioning 2.0.0 is from version 0.1.6+
 
 - [x] (Changed) _Front-end_ Upgrade packages to React 19 (PR #2181)
 - [x] (Changed) _Back-end_ Upgrade to .NET 8 - SDK 8.0.410 (Runtime: 8.0.16) (PR #2196)
+- [x] (Added) _Back-end_ Add Camera Body Serial to FileIndexModel (PR #2206)
 
 ### Breaking changes in 0.7.0-beta.0-2